requirements.txt
pyTelegramBotAPI
python-dotenv
watchdog
python-docx
docxtpl